+++ test.py
@@ -1,0 +1,4 @@
+from src.tests.manual import cli
+
+if __name__ == '__main__':
+    cli()
